fix(handoff): complete H1/H2/H3/H4/H5/M8/N7 production fixes; add 7 missing tests
The handoff claimed 7 fixes were already applied. This commit:
- H1: Move HF model family imports to function-local in model_factory.py
- H2: Stream train_loader with iter() + peek (was: list(train_loader))
- H3: Rescale partial-tail loss by grad_accum_steps/window_pos
- H4: Add lazy _taichi_ops_available() probe in gdn2/block.py
- H5: Add d_k read-back term -sum_d S*omega_w_eff*r to delta.py bwd
  Also fix setup_context signature for torch 2.9+ API
- M8: Raise ValueError on fp16+CPU at TrainingLoop.__init__
- N7: Set optimizer._opt_called=True before base_scheduler.step()

Tests added in tests/test_handoff_first_pass.py: 530 passed.

+++ tests/models/test_training.py
@@ -89,7 +89,7 @@ def test_when_training_loop_run_then_advances_global_steps_and_optimizes() -> No
     )
 
     config = TrainConfig(
-        max_steps=4, grad_accum_steps=2, max_grad_norm=1.0, mixed_precision="fp16"
+        max_steps=4, grad_accum_steps=2, max_grad_norm=1.0, mixed_precision="bf16"
     )
     loop = TrainingLoop(model, optimizer, scheduler, config)
